+++ histoprep/__init__.py
@@ -1,4 +1,6 @@
-"""HistoPrep: Preprocessing large medical images for machine learning made easy!"""
+"""HistoPrep: Preprocessing large medical images for machine learning made easy!"""  # noqa
 
-from . import backend, data, functional, utils
-from ._reader import *
+__all__ = ["SlideReader", "functional", "utils"]
+
+from . import functional, utils
+from ._reader import SlideReader
