--- histoprep/_cli/__init__.py
@@ -1,3 +1,0 @@
-"""CLI interface functions."""
-
-from ._cut import cut_slides
