+++ histoprep/_reader.py
@@ -1,68 +1,80 @@
-from __future__ import annotations
-
-__all__ = ["SlideReader"]
-
 import functools
 import json
+import shutil
 from collections.abc import Callable, Iterator
+from dataclasses import dataclass
 from pathlib import Path
-from typing import Any
+from typing import Any, Optional, Union
 
 import numpy as np
 import polars as pl
+import tqdm
 from PIL import Image
 
 import histoprep.functional as F
-from histoprep._helpers import (
-    close_pool,
-    load_region_data,
-    prepare_output_dir,
-    prepare_worker_pool,
-    read_slide,
-    read_tile,
-    save_regions,
-)
-from histoprep.backend import CziBackend, OpenSlideBackend, PillowBackend
-from histoprep.data import SpotCoordinates, TileCoordinates
+from histoprep._backend import CziBackend, OpenSlideBackend, PillowBackend
+from histoprep._data import SpotCoordinates, TileCoordinates
+from histoprep.functional._concurrent import close_pool, prepare_worker_pool
+from histoprep.functional._level import format_level
+from histoprep.functional._tiles import _multiply_xywh
 
 ERROR_WRONG_TYPE = "Expected '{}' to be of type {}, not {}."
 ERROR_NO_THRESHOLD = "Threshold argument is required to save masks/metrics."
+ERROR_AUTOMATIC_BACKEND = (
+    "Could not automatically assing reader for path: '{}'. Please choose from {}."
+)
+ERROR_BACKEND_NAME = "Backend '{}' does not exist, choose from: {}."
+ERROR_OUTPUT_DIR_IS_FILE = "Output directory exists but it is a file."
+ERROR_CANNOT_OVERWRITE = "Output directory exists, but `overwrite=False`."
+AVAILABLE_BACKENDS = ("PILLOW", "OPENSLIDE", "CZI")
+OPENSLIDE_READABLE_FORMATS = (
+    "svs",
+    "vms",
+    "vmu",
+    "ndpi",
+    "scn",
+    "mrxs",
+    "tiff",
+    "svslide",
+    "tif",
+    "bif",
+)
 
 
 class SlideReader:
-    """Reader class for histological slide images.
-
-    This class wraps around `histoprep.backend` image readers and extends them with
-    functionality for extracting and saving image regions and much more!
-
-    Args:
-        path: Path to slide image.
-        backend: Backend to use for reading slide images. If None, attempts to
-            assing the correct backend based on file extension. Defaults to None.
-    """
+    """Reader class for histological slide images."""
 
     def __init__(
         self,
-        path: str | Path,
-        backend: str | OpenSlideBackend | PillowBackend | CziBackend | None = None,
+        path: Union[str, Path],
+        backend: Optional[
+            Union[str, OpenSlideBackend, PillowBackend, CziBackend]
+        ] = None,
     ) -> None:
+        """Initialize `SlideReader` instance.
+
+        Args:
+            path: Path to slide image.
+            backend: Backend to use for reading slide images. If None, attempts to
+                assing the correct backend based on file extension. Defaults to None.
+        """
         super().__init__()
-        self.backend = read_slide(path=path, backend=backend)
+        self._backend = _read_slide(path=path, backend=backend)
 
     @property
     def path(self) -> str:
         """Full slide filepath."""
-        return self.backend.path
+        return self._backend.path
 
     @property
     def name(self) -> str:
         """Slide filename without an extension."""
-        return self.backend.name
+        return self._backend.name
 
     @property
     def suffix(self) -> str:
         """Slide file-extension."""
-        return self.backend.suffix
+        return self._backend.suffix
 
     @property
     def data_bounds(self) -> tuple[int, int, int, int]:
@@ -73,27 +85,27 @@ class SlideReader:
         full image dimensions, but other software (such as `QuPath`) uses the image
         dimensions defined by this data bound.
         """
-        return self.backend.data_bounds
+        return self._backend.data_bounds
 
     @property
     def dimensions(self) -> tuple[int, int]:
         """Image dimensions (height, width) at `level=0`."""
-        return self.backend.dimensions
+        return self._backend.dimensions
 
     @property
     def level_count(self) -> int:
         """Number of slide levels."""
-        return self.backend.level_count
+        return self._backend.level_count
 
     @property
     def level_dimensions(self) -> dict[int, tuple[int, int]]:
         """Image dimensions (height, width) for each level."""
-        return self.backend.level_dimensions
+        return self._backend.level_dimensions
 
     @property
     def level_downsamples(self) -> dict[int, tuple[float, float]]:
         """Image downsample factors (height, width) for each level."""
-        return self.backend.level_downsamples
+        return self._backend.level_downsamples
 
     def read_level(self, level: int) -> np.ndarray:
         """Read full level data.
@@ -107,7 +119,7 @@ class SlideReader:
         Returns:
             Array containing image data from `level`.
         """
-        return self.backend.read_level(level=level)
+        return self._backend.read_level(level=level)
 
     def read_region(
         self, xywh: tuple[int, int, int, int], level: int = 0
@@ -124,7 +136,7 @@ class SlideReader:
         Returns:
             Array containing image data from `xywh`-region.
         """
-        return self.backend.read_region(xywh=xywh, level=level)
+        return self._backend.read_region(xywh=xywh, level=level)
 
     def level_from_max_dimension(self, max_dimension: int = 4096) -> int:
         """Find level with both dimensions less or equal to `max_dimension`.
@@ -135,9 +147,10 @@ class SlideReader:
         Returns:
             Slide level.
         """
-        return F._level_from_max_dimension(
-            max_dimension=max_dimension, level_dimensions=self.level_dimensions
-        )
+        for level, (level_h, level_w) in self.level_dimensions.items():
+            if level_h <= max_dimension and level_w <= max_dimension:
+                return level
+        return list(self.level_dimensions.keys())[-1]
 
     def level_from_dimensions(self, dimensions: tuple[int, int]) -> int:
         """Find level which is closest to `dimensions`.
@@ -148,15 +161,19 @@ class SlideReader:
         Returns:
             Slide level.
         """
-        return F._level_from_dimensions(
-            dimensions=dimensions, level_dimensions=self.level_dimensions
-        )
+        height, width = dimensions
+        available = []
+        distances = []
+        for level, (level_h, level_w) in self.level_dimensions.items():
+            available.append(level)
+            distances.append(abs(level_h - height) + abs(level_w - width))
+        return available[distances.index(min(distances))]
 
     def get_tissue_mask(
         self,
         *,
-        level: int | None = None,
-        threshold: int | None = None,
+        level: Optional[int] = None,
+        threshold: Optional[int] = None,
         multiplier: float = 1.05,
         sigma: float = 0.0,
     ) -> tuple[int, np.ndarray]:
@@ -183,7 +200,7 @@ class SlideReader:
         level = (
             self.level_from_max_dimension()
             if level is None
-            else F._format_level(level, available=list(self.level_dimensions))
+            else format_level(level, available=list(self.level_dimensions))
         )
         return F.get_tissue_mask(
             image=self.read_level(level),
@@ -194,10 +211,10 @@ class SlideReader:
 
     def get_tile_coordinates(
         self,
-        tissue_mask: np.ndarray | None,
+        tissue_mask: Optional[np.ndarray],
         width: int,
         *,
-        height: int | None = None,
+        height: Optional[int] = None,
         overlap: float = 0.0,
         max_background: float = 0.95,
         out_of_bounds: bool = True,
@@ -255,9 +272,9 @@ class SlideReader:
         tissue_mask: np.ndarray,
         *,
         min_area_pixel: int = 10,
-        max_area_pixel: int | None = None,
+        max_area_pixel: Optional[int] = None,
         min_area_relative: float = 0.2,
-        max_area_relative: float | None = 2.0,
+        max_area_relative: Optional[float] = 2.0,
     ) -> SpotCoordinates:
         """Generate tissue microarray spot coordinates.
 
@@ -287,7 +304,7 @@ class SlideReader:
         # Dearray spots.
         spot_info = F.get_spot_coordinates(spot_mask)
         spot_coordinates = [  # upsample to level zero.
-            F._multiply_xywh(x, F.get_downsample(tissue_mask, self.dimensions))
+            _multiply_xywh(x, F.get_downsample(tissue_mask, self.dimensions))
             for x in spot_info.values()
         ]
         return SpotCoordinates(
@@ -328,17 +345,17 @@ class SlideReader:
             )
         else:
             kwargs.update({"coordinates": coordinates})
-        return F.draw_tiles(**kwargs)
+        return F.get_annotated_image(**kwargs)
 
     def yield_regions(
         self,
         coordinates: Iterator[tuple[int, int, int, int]],
         *,
         level: int = 0,
-        transform: Callable[[np.ndarray], Any] | None = None,
+        transform: Optional[Callable[[np.ndarray], Any]] = None,
         num_workers: int = 1,
         return_exception: bool = False,
-    ) -> Iterator[tuple[np.ndarray | Exception | Any, tuple[int, int, int, int]]]:
+    ) -> Iterator[tuple[Union[np.ndarray, Exception, Any], tuple[int, int, int, int]]]:
         """Yield tile images and corresponding xywh coordinates.
 
         Args:
@@ -355,7 +372,7 @@ class SlideReader:
         """
         pool, iterable = prepare_worker_pool(
             worker_fn=functools.partial(
-                read_tile,
+                _read_tile,
                 level=level,
                 transform=transform,
                 return_exception=return_exception,
@@ -409,16 +426,16 @@ class SlideReader:
 
     def save_regions(
         self,
-        parent_dir: str | Path,
+        parent_dir: Union[str, Path],
         coordinates: Iterator[tuple[int, int, int, int]],
         *,
         level: int = 0,
-        threshold: int | None = None,
+        threshold: Optional[int] = None,
         overwrite: bool = False,
         save_metrics: bool = False,
         save_masks: bool = False,
         save_thumbnails: bool = True,
-        thumbnail_level: int | None = None,
+        thumbnail_level: Optional[int] = None,
         image_format: str = "jpeg",
         quality: int = 80,
         use_csv: bool = False,
@@ -464,9 +481,9 @@ class SlideReader:
         """
         if (save_metrics or save_masks) and threshold is None:
             raise ValueError(ERROR_NO_THRESHOLD)
-        level = F._format_level(level, available=list(self.level_dimensions))
+        level = format_level(level, available=list(self.level_dimensions))
         parent_dir = parent_dir if isinstance(parent_dir, Path) else Path(parent_dir)
-        output_dir = prepare_output_dir(parent_dir / self.name, overwrite=overwrite)
+        output_dir = _prepare_output_dir(parent_dir / self.name, overwrite=overwrite)
         image_dir = "spots" if isinstance(coordinates, SpotCoordinates) else "tiles"
         # Save properties.
         if isinstance(coordinates, TileCoordinates):
@@ -492,13 +509,13 @@ class SlideReader:
                 Image.fromarray(255 - 255 * coordinates.tissue_mask).save(
                     output_dir / "thumbnail_tissue.jpeg"
                 )
-        metadata = save_regions(
+        metadata = _save_regions(
             output_dir=output_dir,
             iterable=self.yield_regions(
                 coordinates=coordinates,
                 level=level,
                 transform=functools.partial(
-                    load_region_data,
+                    _load_region_data,
                     save_masks=save_masks,
                     save_metrics=save_metrics,
                     threshold=threshold,
@@ -525,5 +542,196 @@ class SlideReader:
     def __repr__(self) -> str:
         return (
             f"{self.__class__.__name__}(path={self.path}, "
-            f"backend={self.backend.BACKEND_NAME})"
+            f"backend={self._backend.BACKEND_NAME})"
         )
+
+
+@dataclass
+class RegionData:
+    """Dataclass representing data for a slide region."""
+
+    image: np.ndarray
+    mask: Optional[np.ndarray]
+    metrics: dict[str, float]
+
+    def save_data(
+        self,
+        *,
+        image_dir: Path,
+        mask_dir: Path,
+        xywh: tuple[int, int, int, int],
+        quality: int,
+        image_format: str,
+        prefix: Optional[str],
+    ) -> dict[str, float]:
+        """Save image (and mask) and return region metadata."""
+        metadata = dict(zip("xywh", xywh))
+        filename = "x{}_y{}_w{}_h{}".format(*xywh)
+        if prefix is not None:
+            filename = f"{prefix}_{filename}"
+        # Save image.
+        image_path = image_dir / f"{filename}.{image_format}"
+        image_path.parent.mkdir(parents=True, exist_ok=True)
+        Image.fromarray(self.image).save(image_path, quality=quality)
+        metadata["path"] = str(image_path.resolve())
+        # Save mask.
+        if self.mask is not None:
+            mask_path = mask_dir / f"{filename}.png"
+            mask_path.parent.mkdir(parents=True, exist_ok=True)
+            Image.fromarray(self.mask).save(mask_path)
+            metadata["mask_path"] = str(mask_path.resolve())
+        return {**metadata, **self.metrics}
+
+
+def _read_slide(  # noqa
+    path: Union[str, Path], backend: Optional[str] = None
+) -> Union[CziBackend, OpenSlideBackend, PillowBackend]:
+    """Read slide with the requested backend.
+
+    Args:
+        backend: Backend to use for reading slide images. If None, attempts to
+            assing the correct backend based on file extension. Defaults to None.
+
+    Raises:
+        FileNotFoundError: Path does not exist.
+        ValueError: Cannot automatically assign backend for reader.
+        ValueError: Backend name not recognised.
+
+    Returns:
+        Slide reader backend.
+    """
+    if not isinstance(path, Path):
+        path = Path(path)
+    if not path.exists():
+        raise FileNotFoundError(str(path.resolve()))
+    if backend is None:
+        # Based on file-extension.
+        if path.name.endswith(OPENSLIDE_READABLE_FORMATS):
+            return OpenSlideBackend(path)
+        if path.name.endswith(("jpeg", "jpg")):
+            return PillowBackend(path)
+        if path.name.endswith("czi"):
+            return CziBackend(path)
+        raise ValueError(ERROR_AUTOMATIC_BACKEND.format(path, AVAILABLE_BACKENDS))
+    if isinstance(backend, str):
+        # Based on backend argument.
+        if "PIL" in backend.upper():
+            return PillowBackend(path)
+        if "OPEN" in backend.upper():
+            return OpenSlideBackend(path)
+        if "CZI" in backend.upper() or "ZEISS" in backend.upper():
+            return CziBackend(path)
+    if isinstance(
+        backend, (type(CziBackend), type(OpenSlideBackend), type(PillowBackend))
+    ):
+        return backend(path=path)
+    raise ValueError(ERROR_BACKEND_NAME.format(backend, AVAILABLE_BACKENDS))
+
+
+def _read_tile(
+    worker_state: dict,
+    xywh: tuple[int, int, int, int],
+    *,
+    level: int,
+    transform: Optional[Callable[[np.ndarray], Any]],
+    return_exception: bool,
+) -> Union[np.ndarray, Exception, Any]:
+    """Parallisable tile reading function."""
+    reader = worker_state["reader"]
+    try:
+        tile = reader.read_region(xywh=xywh, level=level)
+    except KeyboardInterrupt:
+        raise KeyboardInterrupt from None
+    except Exception as catched_exception:  # noqa
+        if not return_exception:
+            raise catched_exception  # noqa
+        return catched_exception
+    if transform is not None:
+        return transform(tile)
+    return tile
+
+
+def _prepare_output_dir(output_dir: Union[str, Path], *, overwrite: bool) -> Path:
+    """Prepare output directory for saving regions."""
+    if not isinstance(output_dir, Path):
+        output_dir = Path(output_dir)
+    if output_dir.exists():
+        if output_dir.is_file():
+            raise NotADirectoryError(ERROR_OUTPUT_DIR_IS_FILE)
+        if len(list(output_dir.iterdir())) > 0 and not overwrite:
+            raise ValueError(ERROR_CANNOT_OVERWRITE)
+        shutil.rmtree(output_dir)
+    output_dir.mkdir(parents=True, exist_ok=True)
+    return output_dir
+
+
+def _load_region_data(
+    image: np.ndarray,
+    *,
+    save_masks: bool,
+    save_metrics: bool,
+    threshold: Optional[int],
+) -> RegionData:
+    """Helper transform to add tissue mask and image metrics during tile loading."""
+    tissue_mask = None
+    metrics = {}
+    if save_masks or save_metrics:
+        __, tissue_mask = F.get_tissue_mask(image=image, threshold=threshold)
+    if save_metrics:
+        metrics = F.get_image_metrics(image=image, tissue_mask=tissue_mask)
+    return RegionData(
+        image=image, mask=tissue_mask if save_masks else None, metrics=metrics
+    )
+
+
+def _save_regions(
+    output_dir: Path,
+    iterable: Iterator[RegionData, tuple[int, int, int, int]],
+    *,
+    desc: str,
+    total: int,
+    quality: int,
+    image_format: str,
+    image_dir: str,
+    file_prefixes: list[str],
+    verbose: bool,
+) -> pl.DataFrame:
+    """Save region data to output directory.
+
+    Args:
+        output_dir: Output directory.
+        iterable: Iterable yieldin RegionData and xywh-coordinates.
+        desc: For the progress bar.
+        total: For the progress bar.
+        quality: Quality of jpeg-compression.
+        image_format: Image extension.
+        image_dir: Image directory name
+        file_prefixes: List of file prefixes.
+        verbose: Enable progress bar.
+
+    Returns:
+        Polars dataframe with metadata.
+    """
+    progress_bar = tqdm.tqdm(
+        iterable=iterable,
+        desc=desc,
+        disable=not verbose,
+        total=total,
+    )
+    rows = []
+    num_failed = 0
+    for i, (region_data, xywh) in enumerate(progress_bar):
+        if isinstance(region_data, Exception):
+            num_failed += 1
+            progress_bar.set_postfix({"failed": num_failed}, refresh=False)
+        rows.append(
+            region_data.save_data(
+                image_dir=output_dir / image_dir,
+                mask_dir=output_dir / "masks",
+                xywh=xywh,
+                quality=quality,
+                image_format=image_format,
+                prefix=None if file_prefixes is None else file_prefixes[i],
+            )
+        )
+    return pl.DataFrame(rows)
