--- histoprep/data/__init__.py
@@ -1,3 +1,0 @@
-"""Data classes."""
-
-from ._coordinates import *
