+++ histoprep/functional/_metrics.py
@@ -1,7 +1,4 @@
-from __future__ import annotations
-
-__all__ = ["calculate_image_metrics"]
-
+from typing import Union
 
 import cv2
 import numpy as np
@@ -21,8 +18,8 @@ BLACK_PIXEL = 0
 WHITE_PIXEL = 255
 
 
-def calculate_image_metrics(
-    image: Image.Image | np.ndarray,
+def get_image_metrics(
+    image: Union[Image.Image, np.ndarray],
     tissue_mask: np.ndarray,
     quantiles: tuple[float, ...] = DEFAULT_QUANTILES,
     shape: tuple[int, int] = DEFAULT_SHAPE,
