+++ histoprep/utils/__init__.py
@@ -1,3 +1,14 @@
-"""Utility functions and classes for processing metadata."""
+"""Utility functions and classes processing tile images, normalizing stains and creating
+datasets."""
 
-from . import normalize, process, torch
+__all__ = [
+    "MachenkoStainNormalizer",
+    "SlideReaderDataset",
+    "TileImageDataset",
+    "TileMetadata",
+    "VahadaneStainNormalizer",
+]
+
+from ._normalize import MachenkoStainNormalizer, VahadaneStainNormalizer
+from ._process import TileMetadata
+from ._torch import SlideReaderDataset, TileImageDataset
