+++ pyproject.toml
@@ -6,7 +6,7 @@ authors = ["jopo666 <jopo@birdlover.com>"]
 readme = "README.md"
 
 [tool.poetry.scripts]
-HistoPrep = 'histoprep._cli._cut:cut_slides'
+HistoPrep = 'histoprep._cli:cut_slides'
 
 [tool.poetry.dependencies]
 aicspylibczi = "^3"
@@ -28,6 +28,7 @@ pytest = "*"
 ipywidgets = "*"
 torch = "*"
 pytest-console-scripts = "^1.3.1"
+pdoc = "^13"
 
 [build-system]
 requires = ["poetry-core"]
