+++ tests/collage_test.py
